# Conflicts:
#	rm-datasource/src/main/java/io/seata/rm/datasource/DataSourceProxy.java
#	rm-datasource/src/main/resources/META-INF/services/io.seata.rm.datasource.exec.InsertExecutor
#	rm-datasource/src/main/resources/META-INF/services/io.seata.rm.datasource.sql.struct.TableMetaCache
#	rm-datasource/src/main/resources/META-INF/services/io.seata.rm.datasource.undo.UndoExecutorHolder
#	rm-datasource/src/main/resources/META-INF/services/io.seata.rm.datasource.undo.UndoLogManager
#	rm-datasource/src/main/resources/META-INF/services/io.seata.sqlparser.KeywordChecker
#	rm-datasource/src/test/java/io/seata/rm/datasource/DataSourceProxyTest.java
#	sqlparser/seata-sqlparser-druid/src/main/resources/META-INF/services/io.seata.sqlparser.druid.SQLOperateRecognizerHolder

+++ common/src/main/java/io/seata/common/util/BlobUtils.java
@@ -64,7 +64,7 @@ public class BlobUtils {
         }
 
         try {
-            return new String(blob.getBytes((long) 1, (int) blob.length()), Constants.DEFAULT_CHARSET);
+            return new String(blob.getBytes(1, (int) blob.length()), Constants.DEFAULT_CHARSET);
         } catch (Exception e) {
             throw new ShouldNeverHappenException(e);
         }
@@ -100,7 +100,7 @@ public class BlobUtils {
         }
 
         try {
-            return blob.getBytes((long) 1, (int) blob.length());
+            return blob.getBytes(1, (int) blob.length());
         } catch (Exception e) {
             throw new ShouldNeverHappenException(e);
         }
